+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   - Added support for specifying a lobby tag
   - Fixed challenge moon lobbies not showing the "challenge moon" text on the lobby list
   - Replaced the lobby list distance dropdown with a dropdown for sorting the list
+- Steam Lobby List Fixes
+  - Fixed the button to copy a lobby id not showing
 
 ### 1.0.0